+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Features
 
 * Add techdivision/import#146
+* Add techdivision/import#162
 * Switch to latest techdivision/import 15.* version as dependency
 
 # Version 15.0.0